+++ packages/init-yount/CHANGELOG.md
@@ -7,8 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Install yount in script.
+
 ### Fixed
 
+- Script intro message.
 - Schema file path in "Next Steps" message.
 
 ## [1.0.0-beta-1] - 2023-03-31